refactor: extract editor page and use usemapContext
